# Speedup for ef3m in bet sizing