# Description

Increases speed of calculations by making some minor changes to the code to make it compatible with Numba.

## Type of change

- [x] Improvement (modifies existing code to improve speed, usability, readability etc.)

# How Has This Been Tested?

- [x] ```python -m unittest test_bet_sizing.TestBetSizeReserve.test_bet_size_reserve_default```
- [x] ```python -m unittest test_ef3m.py```

**Test Configuration**:
* Manjaro Linux
* Visual Studio Code


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes

# Notes

Below are the timed test results for ```python -m unittest test_bet_sizing.TestBetSizeReserve.test_bet_size_reserve_default``` which should give an idea for the aprx. increase in speed:

Native:
```
Ran 1 test in 79.200s

OK
python -m unittest   80,26s user 0,23s system 99% cpu 1:20,86 total

```

Numba:
```
Ran 1 test in 45.126s

OK
python -m unittest   46,31s user 0,28s system 99% cpu 46,812 total
```

I also ran ```python -m unittest test_bet_sizing.TestBetSizeReserve.test_bet_size_reserve_default``` modified to use variant 1:

Native:
```
Ran 1 test in 5876.010s

OK
python -m unittest   5862,62s user 2,21s system 99% cpu 1:37:57,66 total
```

Numba:
```
Ran 1 test in 1498.253s

OK
python -m unittest   1495,46s user 0,55s system 99% cpu 24:59,90 total
```

When running ```python -m unittest test_ef3m.py``` the Numba version is actually slower. This is probably due to the overhead of the Numba JIT compiler offsetting any speed gains. I am assuming the `test_ef3m` tests were written with a small data size in mind and the previous tests are more representative for real world usage:

Native:
```
Ran 29 tests in 5.213s

OK
python -m unittest test_ef3m.py  6,60s user 0,16s system 98% cpu 6,849 total
```

Numba:
```
Ran 29 tests in 8.635s

OK
python -m unittest test_ef3m.py  10,05s user 0,19s system 98% cpu 10,369 total
```