@Jackal08 2020-04-24
I can't tell you how happy I am that you are speeding this module up! It's largely the reason our build scripts take so long! This will make a big difference in our wait time for builds. Thank you! Ping me when the tests are passing and I will gladly do the review.